requirements.txt
asgiref==3.8.1
black==24.8.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Django==4.2
django-filter==24.3
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==29.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pycparser==2.22
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.2